Add required = 'y in template' option to FieldMapping (#59)
This is used by special record types for relationships.

These record types do not actually need object/subject types indicated
for successful mapping. They do need to be specified in order to look
up the CSID of the subject/object, however. So they must be in the CSV
template (and should be marked as required there), but are not
required after the data has been transformed for mapping (i.e.
terms/obj numbers replaced with CSIDs)

Also, whatever is listed as "identifier_field" in the RecordMapper
will be treated as a required field for mapping. Since users should
not have to specify an id for a relationship record, I've set this to
subjectCsid in all cases, though that is not actually correct.